backend/requirements.txt
django<6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
gunicorn
whitenoise
pillow
django-environ
dj-database-url
